requirements.txt
amqp==2.6.1
asn1crypto==0.24.0
astroid==2.5.3
attrs==20.3.0
autopep8==1.5.7
billiard==3.5.0.5
celery==4.2.2
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
cryptography==3.4.7
Django==2.2.15
django-auditlog==0.4.7
django-cors-headers==3.7.0
django-crontab==0.7.1
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-jsonfield==1.4.1
django-test-without-migrations==0.6
djangorestframework==3.10.0
djangorestframework-simplejwt==4.4.0
drf-spectacular==0.15.1
fakeredis==1.5.0
flake8==3.9.1
gunicorn==20.1.0
httplib2==0.20.2
idna==2.8
importlib-metadata==3.10.1
inflection==0.5.1
isort==5.8.0
Jinja2==3.0.1
jsonschema==3.2.0
kombu==4.3.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.17.3
pycodestyle==2.7.0
pycparser==2.19
pyflakes==2.3.1
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.7.4
pyparsing==3.0.6
pyrsistent==0.17.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.6.0
python-decouple==3.3
pytz==2018.7
PyYAML==5.3.1
redis==3.2.1
requests==2.21.0
sentry-sdk==0.12.3
six==1.12.0
sortedcontainers==2.2.2
sqlparse==0.3.0
toml==0.10.1
twilio==6.0.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.24.1
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
fcm-django==1.0.6
django-request-logging==0.7.5
sigfig